% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/venndia.R
\name{circle}
\alias{circle}
\title{circle}
\usage{
circle(x, y, r, ...)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}

\item{r}{radius}

\item{...}{}
}
\value{
A graphic of a circle
}
\description{
A function to draw circles.
}
\examples{
# No example
}

