% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getProbeID.R
\name{getProbeID}
\alias{getProbeID}
\title{getProbeID}
\usage{
getProbeID(gene)
}
\arguments{
\item{gene}{A string or vector of strings specifying a gene or list of genes to convert to Affymetrix probe_id format.  Can take AGI identifiers (eg At1g12345), Affymetrix probe_ids (end in _at), or short names such as ACT1, UBQ11 (not recommended since multiple genes can share the same short name).}
}
\value{
A string or vector of strings representing the Affymetrix probe_id unique identifiers of the gene(s) input (where multiple probe_ids exist for a single gene, the first will be returned)
}
\description{
Input is a gene name: short name (GID1b), AT number (At5g60100), or probe_id. Input is case-insensitive unless you input the probe_ids (must end in "_at"). Output is the probe_id for that gene on ATH1 Arabidopsis microarray.
}
\examples{
# Get the probe_ids of single genes
getProbeID("At5g60100")
getProbeID("SLY1")
getProbeID("245790_at")

# Get probe_ids from a list of genes
x <- c("AT4G25420", "AT5G51810", "AT5G07200", "AT1G30040","AT4G18350", "AT4G19170", "AT3G24220", "AT1G78390")
getProbeID(x)
}

