% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TAGGITontology.R
\name{TAGGITontology}
\alias{TAGGITontology}
\title{TAGGITontology}
\usage{
TAGGITontology(geneList, useSearchTerms = TRUE,
  outputFileName = "TAGGITontologyHits.xlsx", annotationFile = "default",
  taggitAGIs = "default", taggitSearchTerms = "default")
}
\arguments{
\item{geneList}{A vector of genes as AGI identiers or Affymetrix probe_ids.}

\item{useSearchTerms}{Set to \code{FALSE} to prevent search term searches in gene descriptions (faster, but results will be based only on genes specified in \code{taggitAGIs} file)}

\item{outputFileName}{Path to output excel file containing lists of hits with descriptions.  File is output into the current working directory as \code{"TAGGITontologyHits.xlsx"} by default.  Set to \code{"none"} for no output.}

\item{annotationFile}{Use default unless you wish to provide your own updated \code{myAnnot} file. See description for instructions for creating this file.}

\item{taggitAGIs}{Generally use default.  Can also pass the path of your own taggitAGIs.tsv (tab-separated) file defining specific AGI identifiers for each TAGGIT term.}

\item{taggitSearchTerms}{Generally use default.  Can also pass the path of your own taggitSearchTerms.tsv (tab-separated) file defining TAGGIT term-specific search terms to search for within gene descriptions.}
}
\value{
A dataframe containing lists of hits for each TAGGIT term.  Also can save an excel spreadsheet containing these hits with descriptions separated by TAGGIT term.
}
\description{
Original TAGGIT paper: (Carrera et al., 2007)
This function is hardcoded for use with Arabidopsis ATH1 microarray data
To use with data from another organism, you will need make a few edits.
}
\examples{
# TBD
}

