% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TAGGITplot.R
\name{TAGGITplot}
\alias{TAGGITplot}
\title{TAGGITplot}
\usage{
TAGGITplot(geneListA, geneListB, ontologyCountsA, ontologyCountsB,
  A = "UPreg", B = "DOWNreg", labA = A, labB = B,
  title = "TAGGIT ontology")
}
\arguments{
\item{geneListA}{A vector of genes as AGI identiers or Affymetrix probe_ids.}

\item{geneListB}{Another vector of genes as AGI identiers or Affymetrix probe_ids.}

\item{ontologyCountsA}{The dataframe resulting from a \code{TAGGITontology()} call on \code{geneListA}.}

\item{ontologyCountsB}{The dataframe resulting from a \code{TAGGITontology()} call on \code{geneListB}.}

\item{A}{UPreg by default (colors will be red/blue), change if not comparing up- to down-regulated geneLists and colors will also change.}

\item{B}{DOWNreg by default (colors will be red/blue), change if not comparing up- to down-regulated geneLists and colors will also change.}

\item{labA}{Label for A (defaults to whatever was passed to \code{A}).}

\item{labB}{Label for B (defaults to whatever was passed to \code{B}).}

\item{title}{Default is \code{"TAGGIT ontology"}.  Set to \code{""} for no title.}
}
\value{
A plot
}
\description{
Takes a TAGGITontology object and plots it using ggplot2
}
\examples{
# TBD
}

