% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TAGGITontology.R
\name{TAGGITontology}
\alias{TAGGITontology}
\title{TAGGITontology}
\usage{
TAGGITontology(geneList, useSearchTerms = TRUE,
  outputFileName = "TAGGITontologyHits.xlsx", annotationFile = "default",
  taggitAGIs = "default", taggitSearchTerms = "default")
}
\arguments{
\item{geneList}{A vector of genes as AGI identiers or Affymetrix probe_ids.}

\item{useSearchTerms}{Set to \code{FALSE} to prevent search term searches in gene descriptions (faster, but results will be based only on genes specified in \code{taggitAGIs} file)}

\item{outputFileName}{Path to output excel file containing lists of hits with descriptions.  File is output into the current working directory as \code{"TAGGITontologyHits.xlsx"} by default.  Set to \code{"none"} for no output.}

\item{annotationFile}{Use default unless you wish to provide your own updated \code{myAnnot} file. See description for instructions for creating this file.}

\item{taggitAGIs}{Generally use default.  Can also pass the path of your own taggitAGIs.tsv (tab-separated) file defining specific AGI identifiers for each TAGGIT term.}

\item{taggitSearchTerms}{Generally use default.  Can also pass the path of your own taggitSearchTerms.tsv (tab-separated) file defining TAGGIT term-specific search terms to search for within gene descriptions.}
}
\value{
A dataframe containing lists of hits for each TAGGIT term.  Also can save an excel spreadsheet containing these hits with descriptions separated by TAGGIT term.
}
\description{
Original TAGGIT paper: (Carrera et al., 2007)
This function is hardcoded for use with Arabidopsis ATH1 microarray data
To use with data from another organism, you will need make a few edits.
}
\examples{
# Example data is provided, use the following code to load it
data("GeneSetdata") # loads GeneSet_UP and GeneSet_DOWN objects which contain example data.

## Make a list containing two vectors: 1) up-regulated genes, and 2) down-regulated gene
GeneSet <- list()
GeneSet$UP <- GeneSet_UP # A vector of upregulated genes
GeneSet$DN <- GeneSet_DOWN # A vector of down-regulated genes

## Create TAGGITontology objects
# (may take a few minutes, but the built-in progress bar will keep you informed on the progress)
# Returns dataframe for plotting and outputs an excel sheet of hits to the working directory.
GeneSet_TAGGIT_UP <- TAGGITontology(GeneSet$UP, outputFileName = "TAGGITontologyHits_UP.xlsx")
GeneSet_TAGGIT_DN <- TAGGITontology(GeneSet$DN, outputFileName = "TAGGITontologyHits_DN.xlsx")

## Plot the results of the TAGGIT analysis using ggplot2 via the TAGGITplot function
# Comparing UP and DOWN regulated genesets
TAGGITplot(GeneSet$UP, GeneSet$DN, GeneSet_TAGGIT_UP, GeneSet_TAGGIT_DN, title = "")
## Export the image: for best results export in EPS (vector) format
## For output like Figure 4 in the publication, export at 400x511 resolution
}

